+++ src/Service/Proxy/Port/ProxyService.php
@@ -2,14 +2,18 @@
 
 namespace FluxIliasRestApi\Service\Proxy\Port;
 
+use FluxIliasRestApi\Adapter\Proxy\WebProxyMapDto;
 use FluxIliasRestApi\Adapter\User\UserDto;
 use FluxIliasRestApi\Service\ConfigForm\Port\ConfigFormService;
+use FluxIliasRestApi\Service\Constants\Port\ConstantsService;
 use FluxIliasRestApi\Service\FluxIliasRestObject\Port\FluxIliasRestObjectService;
 use FluxIliasRestApi\Service\Proxy\Command\GetWebProxyCommand;
 use FluxIliasRestApi\Service\Proxy\Command\GetWebProxyMenuItemsCommand;
 use FluxIliasRestApi\Service\Proxy\Command\HandleIliasGotoCommand;
 use FluxIliasRestApi\Service\Proxy\Command\HandleIliasRedirectCommand;
+use FluxIliasRestApi\Service\Proxy\Command\IsWebProxyMenuVisibleCommand;
 use FluxIliasRestApi\Service\ProxyConfig\Port\ProxyConfigService;
+use FluxIliasRestApi\Service\UserRole\Port\UserRoleService;
 use FluxRestApi\Adapter\Api\RestApi;
 use ilGlobalTemplateInterface;
 use ILIAS\DI\Container;
@@ -24,6 +28,8 @@ class ProxyService
         private readonly ConfigFormService $config_form_service,
         private readonly ProxyConfigService $proxy_config_service,
         private readonly FluxIliasRestObjectService $flux_ilias_rest_object_service,
+        private readonly ConstantsService $constants_service,
+        private readonly UserRoleService $user_role_service,
         private readonly Container $ilias_dic
     ) {
 
@@ -35,6 +41,8 @@ class ProxyService
         ConfigFormService $config_form_service,
         ProxyConfigService $proxy_config_service,
         FluxIliasRestObjectService $flux_ilias_rest_object_service,
+        ConstantsService $constants_service,
+        UserRoleService $user_role_service,
         Container $ilias_dic
     ) : static {
         return new static(
@@ -42,6 +50,8 @@ class ProxyService
             $config_form_service,
             $proxy_config_service,
             $flux_ilias_rest_object_service,
+            $constants_service,
+            $user_role_service,
             $ilias_dic
         );
     }
@@ -79,6 +89,7 @@ class ProxyService
     {
         return GetWebProxyMenuItemsCommand::new(
             $this->proxy_config_service,
+            $this,
             $this->ilias_dic
         )
             ->getWebProxyMenuItems(
@@ -112,6 +123,19 @@ class ProxyService
         )
             ->handleIliasRedirect(
                 $url
+            );
+    }
+
+
+    public function isWebProxyMenuVisible(WebProxyMapDto $web_proxy_map, ?UserDto $user) : bool
+    {
+        return IsWebProxyMenuVisibleCommand::new(
+            $this->constants_service,
+            $this->user_role_service
+        )
+            ->isWebProxyMenuVisible(
+                $web_proxy_map,
+                $user
             );
     }
 }